% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{abrir_nova_turma}
\alias{abrir_nova_turma}
\title{Title}
\usage{
abrir_nova_turma(overwrite = FALSE, output_file = NULL)
}
\arguments{
\item{overwrite}{}

\item{output_file}{}
}
\value{

}
\description{
Title
}
