% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{add_interventions}
\alias{add_interventions}
\title{Add all interventions}
\usage{
add_interventions(p, interventions, resistance, irs_adjust)
}
\arguments{
\item{p}{parameter list}

\item{interventions}{site intervention inputs}

\item{resistance}{pyrethroid resistance data for ITN efficacy calculations}

\item{irs_adjust}{scalar adjustment factor to IRS coverage inputs to account for observed discrepancies between trial and real-world implementation}
}
\value{
modified parameter list
}
\description{
Add all interventions
}
