% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs.R
\name{add_irs}
\alias{add_irs}
\title{Add IRS}
\usage{
add_irs(p, irs, irs_adjust = 1)
}
\arguments{
\item{p}{parameter list}

\item{irs}{interventions irs object}

\item{irs_adjust}{Operational downscaling of irs coverage}
}
\value{
modified parameter list
}
\description{
Add IRS
}
