% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itns.R
\name{add_itns}
\alias{add_itns}
\title{Add ITNs}
\usage{
add_itns(p, itn, resistance)
}
\arguments{
\item{p}{parameter list}

\item{itn}{interventions ITN object}

\item{resistance}{vectors resistance object}
}
\value{
modified parameter list
}
\description{
Add ITNs
}
