% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{add_rtss}
\alias{add_rtss}
\title{Add RTS,S}
\usage{
add_rtss(p, interventions)
}
\arguments{
\item{p}{parameter list}

\item{interventions}{site intervention inputs}
}
\value{
modified parameter list
}
\description{
Add RTS,S
}
