% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{add_stephensi_invasion}
\alias{add_stephensi_invasion}
\title{Add LSM}
\usage{
add_stephensi_invasion(p, interventions)
}
\arguments{
\item{p}{parameter list}

\item{interventions}{site intervention inputs}
}
\value{
modified parameter list
}
\description{
Add LSM
}
