% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\name{add_treatment}
\alias{add_treatment}
\title{Add treatment}
\usage{
add_treatment(p, treatment)
}
\arguments{
\item{p}{parameter list}

\item{treatment}{treatment intervention data including coverage and drug type proportions}
}
\value{
modified parameter list
}
\description{
Add treatment
}
