% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccines.R
\name{add_vaccine}
\alias{add_vaccine}
\title{Add vaccine}
\usage{
add_vaccine(p, vaccine)
}
\arguments{
\item{p}{parameter list}

\item{vaccine}{vaccine intervention data including coverage, delivery mode, and booster schedules}
}
\value{
modified parameter list
}
\description{
Add vaccine
}
