% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm.R
\name{adjust_carrying_capacity}
\alias{adjust_carrying_capacity}
\title{Adjust carrying capacity}
\usage{
adjust_carrying_capacity(p, lsm)
}
\arguments{
\item{p}{parameter list}
}
\value{
modified parameter list
}
\description{
Adjust carrying capacity
}
