% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{available_sites}
\alias{available_sites}
\title{View available sites and versions on the server}
\usage{
available_sites()
}
\value{
Version table.
}
\description{
Queries the server for available site files. This could be of use if searching
for older site file versions.
}
