% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge}
\alias{badge}
\title{Create a data source badge}
\usage{
badge(label, colour, link)
}
\arguments{
\item{label}{Text to display in the badge}

\item{colour}{Background color (hex code or CSS color name)}

\item{link}{URL or anchor link (e.g., "https://mrc-ide.github.io/site/articles/data-sources.html#un")}
}
\value{
HTML string for the badge
}
\description{
Create a data source badge
}
