% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{build_legend_data}
\alias{build_legend_data}
\title{Build legend data from present plot elements}
\usage{
build_legend_data(elements, colours, labels)
}
\arguments{
\item{elements}{Character vector of element names present in the plot.}

\item{colours}{Named character vector of colours.}

\item{labels}{Named character vector of labels.}
}
\value{
Data frame with columns: element, colour, label.
}
\description{
Build legend data from present plot elements
}
\keyword{internal}
