% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_site.R
\name{create_example_demography}
\alias{create_example_demography}
\title{Create example demography inputs}
\usage{
create_example_demography()
}
\value{
A data frame with columns:
  name, year, age_lower, age_upper, adjusted_mortality_rates
}
\description{
Creates a demography list for testing
}
