% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_site.R
\name{create_example_resistance}
\alias{create_example_resistance}
\title{Create example resistance data}
\usage{
create_example_resistance()
}
\value{
A data frame with columns: name, year, pyrethroid_resistance
}
\description{
Creates a simple pyrethroid resistance data frame for testing with resistance
levels varying over time.
}
