% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_site.R
\name{create_example_seasonality_parameters}
\alias{create_example_seasonality_parameters}
\title{Create example seasonality inputs}
\usage{
create_example_seasonality_parameters()
}
\value{
A data frame with columns:
  name, g0, g1, g2, g3, h1, h2, h3
}
\description{
Creates a seasonality list for testing
}
