% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_site.R
\name{create_example_site}
\alias{create_example_site}
\title{Create a skeleton example site}
\usage{
create_example_site(itn_use = c(0.1, 0.2, 0.5))
}
\arguments{
\item{itn_use}{Numeric vector. ITN use values to pass to
`create_example_interventions()` (default: c(0.1, 0.2, 0.5)).}
}
\value{
Skeleton example single site
}
\description{
Create a skeleton example site
}
