% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_site.R
\name{create_example_vector_species}
\alias{create_example_vector_species}
\title{Create example vector inputs}
\usage{
create_example_vector_species()
}
\value{
A data frame with columns:
  name, species, prop, blood_meal_rates, foraging time, Q0, phi_bednets, phi_indoors, mum
}
\description{
Creates a vectors list for testing
}
