% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{default_intervention_colours}
\alias{default_intervention_colours}
\title{Default colour palette for intervention plot elements}
\usage{
default_intervention_colours()
}
\value{
Named character vector of colours
}
\description{
Default colour palette for intervention plot elements
}
