% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{default_intervention_labels}
\alias{default_intervention_labels}
\title{Default labels for intervention plot elements}
\usage{
default_intervention_labels()
}
\value{
Named character vector of labels
}
\description{
Default labels for intervention plot elements
}
