% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_site.R
\name{filter_recursive}
\alias{filter_recursive}
\title{Recursively filter site file elements}
\usage{
filter_recursive(x, site_filter)
}
\arguments{
\item{x}{Site file element (can be data.frame, list, or other)}

\item{site_filter}{data frame to filter by}
}
\value{
Filtered element
}
\description{
Recursively filter site file elements
}
