% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irs_efficacy}
\alias{irs_efficacy}
\title{IRS insecticide efficacy parameters}
\format{
A data frame with 4 rows and 7 columns:
\describe{
  \item{insecticide}{Character. Name of the insecticide (ddt, actellic, bendiocarb, sumishield)}
  \item{ls_theta}{Numeric. Theta parameter for mosquito repelling effect}
  \item{ls_gamma}{Numeric. Gamma parameter for mosquito repelling effect}
  \item{ks_theta}{Numeric. Theta parameter for mosquito killing effect}
  \item{ks_gamma}{Numeric. Gamma parameter for mosquito killing effect}
  \item{ms_theta}{Numeric. Theta parameter for successful feeding reduction}
  \item{ms_gamma}{Numeric. Gamma parameter for successful feeding reduction}
}
}
\source{
Parameterized from field trial data for malariasimulation
}
\usage{
irs_efficacy
}
\description{
A dataset containing efficacy parameters for different Indoor Residual Spraying (IRS)
insecticides. These parameters are used by malariasimulation to model the impact of
IRS interventions on mosquito populations.
}
\keyword{datasets}
