% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{known_vector_species}
\alias{known_vector_species}
\title{Known malaria vector species}
\usage{
known_vector_species()
}
\value{
Character vector of all known vector species names.
}
\description{
Known malaria vector species
}
\keyword{internal}
