% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{make_carrying_capacity_matrix}
\alias{make_carrying_capacity_matrix}
\title{Add LSM}
\usage{
make_carrying_capacity_matrix(p, interventions)
}
\arguments{
\item{p}{parameter list}

\item{interventions}{site intervention inputs}
}
\value{
modified parameter list
}
\description{
Add LSM
}
