% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_site.R
\name{match}
\alias{match}
\title{Matched join}
\usage{
match(x, y)
}
\arguments{
\item{x}{Site file element}

\item{y}{Data.frame to match for}
}
\value{
Site file element filtered by y
}
\description{
Matched join
}
