% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_site.R
\name{match_by_names}
\alias{match_by_names}
\title{Matched join}
\usage{
match_by_names(x, y)
}
\arguments{
\item{x}{Site file element}

\item{y}{data frame to match for}
}
\value{
Site file element filtered by y
}
\description{
Matched join
}
