% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_drive.R
\name{pegar_carga_horaria}
\alias{pegar_carga_horaria}
\title{Pega e formata a carga horária do curso}
\usage{
pegar_carga_horaria(id)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Pega e formata a carga horária do curso
}
