% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_drive.R
\name{pegar_info_curso}
\alias{pegar_info_curso}
\title{Pega a informação de algum curso}
\usage{
pegar_info_curso(id, info)
}
\arguments{
\item{id}{}

\item{info}{}
}
\value{

}
\description{
Pega a informação de algum curso
}
