% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_drive.R
\name{pegar_info_turma}
\alias{pegar_info_turma}
\title{Pega a informação da turma mais recente de um curso}
\usage{
pegar_info_turma(id, info)
}
\arguments{
\item{id}{}

\item{info}{}
}
\value{

}
\description{
Pega a informação da turma mais recente de um curso
}
