% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_drive.R
\name{pegar_professores_turma}
\alias{pegar_professores_turma}
\title{Pega a informação da turma de algum curso}
\usage{
pegar_professores_turma(id)
}
\arguments{
\item{id}{}

\item{info}{}
}
\value{

}
\description{
Pega a informação da turma de algum curso
}
