% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Syphproj.R
\name{plot.SummarySyphilisMen}
\alias{plot.SummarySyphilisMen}
\title{Display prevalence results in chart}
\usage{
\method{plot}{SummarySyphilisMen}(result, pop = "all")
}
\arguments{
\item{result:}{list of class "SyphilisSummaryMen"}

\item{pop:}{string, population type. Should be either "all" or men sub-populations}
}
\description{
Display prevalence results in chart
}
