% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pyrethroid_resistance}
\alias{plot_pyrethroid_resistance}
\title{Plot pyrethroid resistance over time}
\usage{
plot_pyrethroid_resistance(pyrethroid_resistance, title = NULL)
}
\arguments{
\item{pyrethroid_resistance}{Data frame with columns: year,
pyrethroid_resistance (0-1)}

\item{title}{Optional plot title.}
}
\value{
A ggplot2 object
}
\description{
Plot pyrethroid resistance over time
}
