% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_site_interventions}
\alias{plot_site_interventions}
\title{Plot interventions from a site object}
\usage{
plot_site_interventions(site, ...)
}
\arguments{
\item{site}{A site object with interventions and seasonality elements.}

\item{...}{Additional arguments passed to [plot_interventions()].}
}
\value{
A ggplot2 object
}
\description{
Plot interventions from a site object
}
