% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_site_prevalence}
\alias{plot_site_prevalence}
\title{Plot site prevalence over time}
\usage{
plot_site_prevalence(prevalence, title = NULL)
}
\arguments{
\item{prevalence}{Data frame with columns: year, pfpr, pvpr}

\item{title}{Optional plot title.}
}
\value{
A ggplot2 object
}
\description{
Plot site prevalence over time
}
