% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_vector_species}
\alias{plot_vector_species}
\title{Plot vector species characteristics as faceted horizontal bars}
\usage{
plot_vector_species(vector_species, title = NULL)
}
\arguments{
\item{vector_species}{Data frame with columns: species, prop, Q0,
phi_indoors, phi_bednets}

\item{title}{Optional plot title.}
}
\value{
A ggplot2 object
}
\description{
Plot vector species characteristics as faceted horizontal bars
}
