% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prepare_continuous_interventions}
\alias{prepare_continuous_interventions}
\title{Prepare continuous intervention coverage data}
\usage{
prepare_continuous_interventions(interventions, years)
}
\arguments{
\item{interventions}{A list containing treatment, pmc, vaccine, and lsm
sub-lists, each with an implementation data frame.}

\item{years}{Numeric vector of years.}
}
\value{
Data frame with columns: t, element, value.
}
\description{
Prepare continuous intervention coverage data
}
\keyword{internal}
