% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prepare_itn}
\alias{prepare_itn}
\title{Prepare ITN data for intervention plot}
\usage{
prepare_itn(itn_implementation, itn_use, retention_half_life, years)
}
\arguments{
\item{itn_implementation}{Data frame with ITN distribution data including
itn_input_dist and distribution_day_of_year columns.}

\item{itn_use}{Data frame with ITN usage data including itn_use and
usage_day_of_year columns.}

\item{retention_half_life}{Numeric net retention half-life in days.}

\item{years}{Numeric vector of years.}
}
\value{
A list with elements: model_use, itn_use, itn_input_dist.
}
\description{
Prepare ITN data for intervention plot
}
\keyword{internal}
