% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prepare_rainfall_bars}
\alias{prepare_rainfall_bars}
\title{Prepare rainfall bar data for intervention plot}
\usage{
prepare_rainfall_bars(monthly_rainfall)
}
\arguments{
\item{monthly_rainfall}{Data frame with columns: year, t, rainfall.}
}
\value{
Data frame with columns: t, value, element.
}
\description{
Prepare rainfall bar data for intervention plot
}
\keyword{internal}
