% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prepare_rainfall_profile}
\alias{prepare_rainfall_profile}
\title{Prepare rainfall profile data for intervention plot}
\usage{
prepare_rainfall_profile(fourier_prediction, years, max_rainfall)
}
\arguments{
\item{fourier_prediction}{Data frame with columns: t, profile.}

\item{years}{Numeric vector of years.}

\item{max_rainfall}{Maximum rainfall value for normalisation.}
}
\value{
Data frame with columns: t, value, element.
}
\description{
Prepare rainfall profile data for intervention plot
}
\keyword{internal}
