% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R, R/r21.R
\docType{data}
\name{r21}
\alias{r21}
\title{Median set of parameters for R21}
\format{
A list with 2 variables:
\describe{
  \item{r21_profile}{profile for the initial 3 doses of R21}
  \item{r21_booster_profile}{profile for booster dose of R21}
}

A list with 2 variables:
\describe{
  \item{r21_profile}{profile for the initial 3 doses of R21}
  \item{r21_booster_profile}{profile for booster dose of R21}
}
}
\usage{
r21

r21
}
\description{
A list containing the median set of parameters for the R21 vaccine per Imperial group draft manuscript.

A list containing the median set of parameters for the R21 vaccine per Imperial group draft manuscript.
}
\keyword{datasets}
