% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{site_vector_palette}
\alias{site_vector_palette}
\title{Vector species colour palette}
\usage{
site_vector_palette(n)
}
\arguments{
\item{n}{Number of colours needed.}
}
\value{
A character vector of hex colours
}
\description{
Distinct colour palette for vector species (categorical, not ordered).
}
