% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_drive.R
\name{verificar_turma_aberta}
\alias{verificar_turma_aberta}
\title{Verifica se existe uma turma aberta}
\usage{
verificar_turma_aberta(id)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Verifica se existe uma turma aberta
}
