% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{year_shading_data}
\alias{year_shading_data}
\title{Create year shading data}
\usage{
year_shading_data(years, offset = "continuous", ymin = -Inf, ymax = Inf)
}
\arguments{
\item{years}{Numeric vector of years.}

\item{offset}{"continuous" (xmin=year, xmax=year+1) or "discrete"
(xmin=year-0.5, xmax=year+0.5).}

\item{ymin, ymax}{Y-axis bounds for shading rectangles.}
}
\value{
A data frame with xmin, xmax, ymin, ymax columns.
}
\description{
Create year shading data
}
\keyword{internal}
