% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{year_shading_layer}
\alias{year_shading_layer}
\title{Year shading geom_rect layer}
\usage{
year_shading_layer(shading_data)
}
\arguments{
\item{shading_data}{Data frame from [year_shading_data()].}
}
\value{
A ggplot2 geom_rect layer.
}
\description{
Year shading geom_rect layer
}
\keyword{internal}
