% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{add_interventions}
\alias{add_interventions}
\title{Add all interventions}
\usage{
add_interventions(p, interventions)
}
\arguments{
\item{p}{parameter list}

\item{interventions}{site intervention inputs}
}
\value{
modified parameter list
}
\description{
Add all interventions
}
