% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{adjust_carrying_capacity}
\alias{adjust_carrying_capacity}
\title{Adjust carrying capacity}
\usage{
adjust_carrying_capacity(p, interventions)
}
\arguments{
\item{p}{parameter list}

\item{interventions}{site intervention inputs}
}
\value{
modified parameter list
}
\description{
Adjust carrying capacity
}
