% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_site.R
\name{create_example_site}
\alias{create_example_site}
\title{Create a skeleton example site}
\usage{
create_example_site()
}
\value{
Skeleton example single site
}
\description{
Create a skeleton example site
}
