% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_site}
\alias{fetch_site}
\title{Fetch a site file for a country from the malariaverse sitefile server.}
\usage{
fetch_site(
  iso3c = NULL,
  version = NULL,
  admin_level = NULL,
  urban_rural = NULL,
  id = NULL
)
}
\arguments{
\item{iso3c}{the ISO country code, a scalar character.}

\item{version}{the dataset version, a scalar character.}

\item{admin_level}{a scalar number.}

\item{urban_rural}{a scalar logical.}

\item{id}{a packet ID used to select an exact packet.}
}
\value{
The contents of the site file.
}
\description{
The site file is identified by its country code, and optionally the
admin_level, urban/rural setting and version of the site files. The latest
packet from the server matching these parameters is used.
By default, calling `fetch_site()` with only `iso3c` downloads the latest
available version. For reproducibility and stable workflows, it is strongly
recommended to specify `version`, `admin_level`, and `urban_rural` explicitly,
or to download and store site files locally. Otherwise, downstream pipelines
may break when new site file versions are released.
}
\details{
Alternatively, a packet ID can be specified in order to pick an exact file
set.
}
\examples{
\dontrun{
fetch_site("NGA")
fetch_site("NGA", admin_level = 1)
fetch_site(id = "20240801-062621-6f95851a")
}
}
