% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_parameters.R
\name{site_parameters}
\alias{site_parameters}
\title{Create site parameters}
\usage{
site_parameters(
  interventions,
  demography,
  vectors,
  seasonality,
  min_ages = c(0, 5, 15) * 365,
  parasite = "falciparum",
  eir = NULL,
  draw = NULL,
  overrides = list(),
  start_year,
  end_year,
  irs_adjust
)
}
\arguments{
\item{interventions}{Site intervention inputs}

\item{demography}{Site demography inputs}

\item{vectors}{Site vectors inputs}

\item{seasonality}{Site seasonality inputs}

\item{min_ages}{Lower age bands for incidence and N age outputs}

\item{parasite}{Can be "falciparum" or "vivax" for vivax SMC, RTSS
and PMC are not implemented}

\item{eir}{Site baseline EIR}

\item{draw}{malariasimulation parameter draw. Default NULL is best-fit parameter set}

\item{overrides}{List of malariasimulation default parameter overrides}

\item{start_year}{}
}
\value{
A malariasimulation parameter list
}
\description{
Create site parameters
}
