% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_parameters.R
\name{site_parameters}
\alias{site_parameters}
\title{Create site parameters}
\usage{
site_parameters(
  interventions,
  demography,
  vectors,
  seasonality,
  min_ages = c(0, 5, 15) * 365,
  eir = NULL,
  overrides = list()
)
}
\arguments{
\item{interventions}{Site intervention inputs}

\item{demography}{Site demography inputs}

\item{vectors}{Site vectors inputs}

\item{seasonality}{Site seasonality inputs}

\item{min_ages}{Lower age bands for incidence and N age outputs}

\item{eir}{Site baseline EIR}

\item{overrides}{List of malariasimulation default parameter overrides}
}
\value{
A malariasimulation parameter list
}
\description{
Create site parameters
}
