% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_parameters.R
\name{site_parameters}
\alias{site_parameters}
\title{Create site parameters}
\usage{
site_parameters(
  interventions,
  demography,
  vectors,
  seasonality,
  eir = NULL,
  baseline_year = 2000,
  overrides = list()
)
}
\arguments{
\item{interventions}{Site intervention inputs}

\item{demography}{Site demography inputs}

\item{vectors}{Site vectors inputs}

\item{seasonality}{Site seasonality inputs}

\item{eir}{Site baseline EIR}

\item{baseline_year}{Simulation baseline year}

\item{overrides}{List of malariasimulation default parameter overrides}
}
\value{
A malariasimulation parameter list
}
\description{
Create site parameters
}
