% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_site.R
\name{subset_site}
\alias{subset_site}
\title{Extract a subset from a country site file}
\usage{
subset_site(site, site_filter)
}
\arguments{
\item{site}{Country site file}

\item{site_filter}{Data.frame to filter site file elements by. Filtering will
be conducted on any matched columns for each element.}
}
\value{
Filtered site file
}
\description{
Extract a subset from a country site file
}
